# MiniLoanManagementService/MiniLoanApiSpec/target/src/main/java/com/aspire/api/LoanApplicationApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.16).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.aspire.api;

import com.aspire.model.ErrorResponse;
import com.aspire.model.LoanApplication;
import com.aspire.model.LoanApplicationResponse;
import com.aspire.model.LoanApprovalResponse;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2024-07-27T07:11:26.518236400+05:30[Asia/Calcutta]")
@Api(value = "LoanApplication", description = "the LoanApplication API")
public interface LoanApplicationApi {

    @ApiOperation(value = "Approve a loan application", nickname = "approveLoanApplication", notes = "Approve Loan Application", response = LoanApprovalResponse.class, authorizations = {
        @Authorization(value = "bearerAuth")    }, tags={ "Loan Application", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Loan approved successfully", response = LoanApprovalResponse.class),
        @ApiResponse(code = 404, message = "Loan not found") })
    @RequestMapping(value = "/api/loan/applications/{applicationId}/approve",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<LoanApprovalResponse> approveLoanApplication(@ApiParam(value = "",required=true) @PathVariable("applicationId") Long applicationId
);


    @ApiOperation(value = "Create a new loan application", nickname = "createLoanApplication", notes = "Creates Loan Application", response = LoanApplicationResponse.class, authorizations = {
        @Authorization(value = "bearerAuth")    }, tags={ "Loan Application", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Loan created successfully", response = LoanApplicationResponse.class),
        @ApiResponse(code = 400, message = "Invalid input", response = ErrorResponse.class) })
    @RequestMapping(value = "/api/loan/applications",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<LoanApplicationResponse> createLoanApplication(@ApiParam(value = "" ,required=true )  @Valid @RequestBody LoanApplication body
);

}